+++ crates/rattler_lock/src/parse/models/v7/mod.rs
@@ -1,4 +1,5 @@
 mod conda_package_data;
+pub(crate) mod deserialize;
 mod pypi_package_data;
 mod source_data;
